refactor: centralize environment variable handling and update action import paths
